Enhance(bsp.c): upgrade static local variables from uint16_t to uint32_t to avoid byte misalignment

+++ G431/G431_03_uart/BSP/Src/bsp.c
@@ -42,7 +42,7 @@ void MainProcess(void)
 
 static void PeriodicProcess(void)
 {
-	static uint16_t cnt = 0;
+	static uint32_t cnt = 0;
 	uint8_t keyStatus = KEY_Read();
 	if (keyStatus == 0x00) return;
