+++ G474/G474_07_adc_data_analysis/BSP/Src/bsp.c
@@ -84,7 +84,7 @@ static void PeriodicProcess(void)
 		HAL_TIM_Base_Start(&htim6);
 	}
 	
-	static uint16_t timeCnt = 0;
+	static uint32_t timeCnt = 0;
 	if (++timeCnt < 10) return;
 	timeCnt = 0;
